// src/services/camunda.enum.ts
enum SortBy {
  InstanceId = 'instanceId',
  DefinitionId = 'definitionId',
  DefinitionKey = 'definitionKey',
  DefinitionName = 'definitionName',
  DefinitionVersion = 'definitionVersion',
  BusinessKey = 'businessKey',
  StartTime = 'startTime',
  EndTime = 'endTime',
  Duration = 'duration',
  TenantId = 'tenantId'
}

enum SortOrder {
  Ascending = 'asc',
  Descending = 'desc'
}

export {SortBy, SortOrder}